moving all commands to their own file
